refactor: vector 패키지 수정
