+++ src/main/java/chess/domain/piece/strategy/vector/KingVector.java
@@ -1,4 +1,4 @@
-package chess.domain.piece;
+package chess.domain.piece.strategy.vector;
 
 import java.util.Arrays;
