+++ src/main/java/chess/domain/piece/strategy/vector/KnightVector.java
@@ -1,4 +1,4 @@
-package chess.domain.piece;
+package chess.domain.piece.strategy.vector;
 
 import java.util.Arrays;
