+++ src/test/java/chess/domain/piece/strategy/vector/DirectionVectorTest.java
@@ -1,17 +1,17 @@
-package chess.domain.piece;
+package chess.domain.piece.strategy.vector;
 
 import static chess.domain.board.File.D;
 import static chess.domain.board.File.E;
 import static chess.domain.board.Rank.FOUR;
 import static chess.domain.board.Rank.THREE;
-import static chess.domain.piece.DirectionVector.EAST;
-import static chess.domain.piece.DirectionVector.NORTH;
-import static chess.domain.piece.DirectionVector.NORTHEAST;
-import static chess.domain.piece.DirectionVector.NORTHWEST;
-import static chess.domain.piece.DirectionVector.SOUTH;
-import static chess.domain.piece.DirectionVector.SOUTHEAST;
-import static chess.domain.piece.DirectionVector.SOUTHWEST;
-import static chess.domain.piece.DirectionVector.WEST;
+import static chess.domain.piece.strategy.vector.DirectionVector.EAST;
+import static chess.domain.piece.strategy.vector.DirectionVector.NORTH;
+import static chess.domain.piece.strategy.vector.DirectionVector.NORTHEAST;
+import static chess.domain.piece.strategy.vector.DirectionVector.NORTHWEST;
+import static chess.domain.piece.strategy.vector.DirectionVector.SOUTH;
+import static chess.domain.piece.strategy.vector.DirectionVector.SOUTHEAST;
+import static chess.domain.piece.strategy.vector.DirectionVector.SOUTHWEST;
+import static chess.domain.piece.strategy.vector.DirectionVector.WEST;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
